+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [6.0.0](https://github.com/stellar/stellar-disbursement-platform-frontend/releases/tag/6.0.0) ([diff](https://github.com/stellar/stellar-disbursement-platform-frontend/compare/5.0.0...6.0.0))
+
+### Added
+- Add Wallet History for Stellar Distribution Accounts. [#385](https://github.com/stellar/stellar-disbursement-platform-frontend/pull/385)
+
+### Fixed 
+- Fix message when adding unknown trustline [#388](https://github.com/stellar/stellar-disbursement-platform-frontend/pull/388)
+
+### Security and Dependencies
+- Upgrade React to 19.2.1 to address CVE-2025-66478 and CVE-2025-55182 [#390](https://github.com/stellar/stellar-disbursement-platform-backend/pull/390)
+- Bump the minor-and-patch group across 1 directory with 30 updates. [#386](https://github.com/stellar/stellar-disbursement-platform-frontend/pull/386)
+- Bump the all-actions group across 1 directory with 4 updates. [#383](https://github.com/stellar/stellar-disbursement-platform-frontend/pull/383)
+- Bump js-yaml from 4.1.0 to 4.1.1. [#381](https://github.com/stellar/stellar-disbursement-platform-frontend/pull/381)
+- Bump vite from 7.1.5 to 7.1.11 in the npm_and_yarn group across 1 directory. [#369](https://github.com/stellar/stellar-disbursement-platform-frontend/pull/369)
+
+
 ## [5.0.0](https://github.com/stellar/stellar-disbursement-platform-frontend/releases/tag/5.0.0) ([diff](https://github.com/stellar/stellar-disbursement-platform-frontend/compare/4.1.0...5.0.0))
 
 ### Added